// packages/api/ring-types.ts
type Firmware = string | 'Up To Date'

// eslint-disable-next-line no-shadow
export enum RingDeviceType {
  BaseStation = 'hub.redsky',
  BaseStationPro = 'hub.kili',
  Keypad = 'security-keypad',
  SecurityPanel = 'security-panel',
  ContactSensor = 'sensor.contact',
  MotionSensor = 'sensor.motion',
  FloodFreezeSensor = 'sensor.flood-freeze',
  FreezeSensor = 'sensor.freeze',
  TemperatureSensor = 'sensor.temperature',
  WaterSensor = 'sensor.water',
  TiltSensor = 'sensor.tilt',
  GlassbreakSensor = 'sensor.glassbreak',
  RangeExtender = 'range-extender.zwave',
  ZigbeeAdapter = 'adapter.zigbee',
  AccessCodeVault = 'access-code.vault',
  AccessCode = 'access-code',
  SmokeAlarm = 'alarm.smoke',
  CoAlarm = 'alarm.co',
  SmokeCoListener = 'listener.smoke-co',
  MultiLevelSwitch = 'switch.multilevel',
  Fan = 'switch.multilevel',
  MultiLevelBulb = 'switch.multilevel.bulb',
  Switch = 'switch',
  BeamsMotionSensor = 'motion-sensor.beams',
  BeamsSwitch = 'switch.beams',
  BeamsMultiLevelSwitch = 'switch.multilevel.beams',
  BeamsLightGroupSwitch = 'group.light-group.beams',
  BeamsTransformerSwitch = 'switch.transformer.beams',
  BeamsDevice = 'device.beams',
  RetrofitBridge = 'bridge.flatline',
  RetrofitZone = 'sensor.zone',
  Thermostat = 'temperature-control.thermostat',
  Sensor = 'sensor',
  RingNetAdapter = 'adapter.ringnet',
  CodeVault = 'access-code.vault',
  SecurityAccessCode = 'access-code',
  ZWaveAdapter = 'adapter.zwave',
  ZWaveExtender = 'range-extender.zwave',
  PanicButton = 'security-panic',
  UnknownZWave = 'unknown.zwave',
}
// ...
export type AlarmMode = 'all' | 'some' | 'none'
export type ThermostatMode = 'heat' | 'cool' | 'off' | 'aux'
export type MessageType =
  | 'RoomGetList'
  | 'SessionInfo'
  | 'DeviceInfoDocGetList'
  | 'DeviceInfoSet'
export type MessageDataType =
  | 'RoomListV2Type'
  | 'SessionInfoType'
  | 'DeviceInfoDocType'
  | 'DeviceInfoSetType'
  | 'HubDisconnectionEventType'

export interface SocketIoMessage {
  msg: MessageType
  datatype: MessageDataType
  src: string
  body: any[]
}

export type AssetKind =
  | 'base_station_v1'
  | 'beams_bridge_v1'
  | 'floodlight_v2'
  | string

export interface AssetSession {
  assetUuid: string
  connectionStatus: 'unknown' | 'cell-backup' | 'online'
  doorbotId: number
  kind: AssetKind
  sessionId: number
}

export type AlarmState =
  | 'burglar-alarm' // Ring is Alarming
  | 'entry-delay' // Alarm will sound in ${timeLeft} seconds
  | 'fire-alarm' // Alarming - Smoke
  | 'co-alarm' // Alarming - CO
  | 'panic' // Panic Triggered
  | 'user-verified-burglar-alarm' // Alarming - User Verified Police
  | 'user-verified-co-or-fire-alarm' // Alarming - User Verified Smoke or CO
  | 'burglar-accelerated-alarm' // Alarming - Police Response Requested
  | 'fire-accelerated-alarm' // Alarming - Fire Department Response Requested

export const allAlarmStates: AlarmState[] = [
  'burglar-alarm',
  'entry-delay',
  'fire-alarm',
  'co-alarm',
  'panic',
  'user-verified-burglar-alarm',
  'user-verified-co-or-fire-alarm',
  'burglar-accelerated-alarm',
  'fire-accelerated-alarm',
]

export interface RingDeviceData {
  zid: string
  name: string
  deviceType: RingDeviceType
  categoryId: number
  batteryLevel?: number
  batteryStatus: 'full' | 'charged' | 'ok' | 'low' | 'none' | 'charging'
  batteryBackup?: 'charged' | 'charging' | 'inUse'
  acStatus?: 'error' | 'ok'
  manufacturerName?: string
  serialNumber?: string
  tamperStatus: 'ok' | 'tamper'
  faulted?: boolean
  locked?: 'jammed' | 'locked' | 'unlocked' | 'unknown'
  roomId?: number
  volume?: number
  mode?: AlarmMode | ThermostatMode
  transitionDelayEndTimestamp?: number | null
  alarmInfo?: {
    state: AlarmState
    faultedDevices?: string[]
    timestamp?: number
    uuid?: string
  }
  siren?: { state: 'on' | 'off' }
  alarmStatus?: 'active'
  co?: { alarmStatus?: 'active' }
  smoke?: { alarmStatus?: 'active' }
  flood?: { faulted?: boolean }
  freeze?: { faulted?: boolean }
  motionStatus?: 'clear' | 'faulted'
  groupId?: string
  tags: ('hidden' | 'sleepy' | 'ota-lock' | 'scanned' | 'kitted' | string)[]

  // switch
  on?: boolean
  // switch.multilevel
  level?: number // 0 - 1
  hs?: {
    hue?: number // 0 - 1
    sat?: number // 0 - 1
  }
  ct?: number // 0 - 1
  // Retrofit sensor.zone
  status?: 'enabled' | 'disabled'
  parentZid?: string
  rootDevice?: string
  relToParentZid?: string // '1' - '8'
  //sensor.temperature
  celsius?: number // no F provided, just celsius
  faultHigh?: number
  faultLow?: number
  // temperature-control.thermostat
  // Related: 'mode?: ThermostatMode' (above)
  setPoint?: number
  setPointMax?: number
  setPointMin?: number
  basicValue?: number // unknown.zwave - 0 for off, 255 for on
  componentDevices?: {
    rel: string
    zid: string
  }[]
  // switch.multilevel.beam
  motionSensorEnabled?: boolean
  // security-keypad
  brightness?: number // 0 - 1
}

export const deviceTypesWithVolume = [
  RingDeviceType.BaseStation,
  RingDeviceType.Keypad,
]

export interface BaseStation {
  address: string
  alerts: any[]
  description: string
  device_id: string
  features: null
  firmware_version: Firmware
  id: number
  kind: string
  latitude: number
  location_id: string
  longitude: number
  owned: boolean
  owner?: {
    id: number
    email: string
    first_name: string
    last_name: string
  }
  ring_id: null
  settings: null
  stolen: boolean
  time_zone: string
}
// ...
export interface LocationAddress {
  address1: string
  address2: string
  cross_street: string
  city: string
  state: string
  timezone: string
  zip_code: string
}

export interface UserLocation {
  address: LocationAddress
  created_at: string
  geo_coordinates: { latitude: string; longitude: string }
  geo_service_verified: 'address_only' | string
  location_id: string
  name: string
  owner_id: number
  updated_at: string
  user_verified: boolean
}

export interface TicketAsset {
  doorbotId: number
  kind: AssetKind
  onBattery: boolean
  status: 'online' | 'offline'
  uuid: string
}
// ...
export type DingKind =
  | 'motion'
  | 'ding'
  | 'on_demand' // Live View
  | 'alarm' // Linked Event - Alarm
  | 'on_demand_link' // Linked Event - Motion
  | 'door_activity' // knock
  | 'key_access'
  | 'DELETED_FOOTAGE'
  | 'OFFLINE_FOOTAGE'
  | 'OFFLINE_MOTION'

export interface CameraEvent {
  created_at: string
  cv_properties: {
    detection_type: null | any
    person_detected: null | any
    stream_broken: null | any
  }
  ding_id: number
  ding_id_str: string
  doorbot_id: number
  favorite: boolean
  kind: DingKind
  recorded: false
  recording_status: 'ready' | 'audio_ready'
  state: 'timed_out' | 'completed'
}

// timed_out + ding === Missed Ring
// completed === Answered

export interface CameraEventResponse {
  events: CameraEvent[]
  meta: { pagination_key: string }
}

export interface CameraEventOptions {
  limit?: number
  kind?: DingKind
  state?: 'missed' | 'accepted' | 'person_detected'
  favorites?: boolean
  olderThanId?: string // alias for pagination_key
  pagination_key?: string
}
// ...
export interface HistoryOptions {
  limit?: number
  offset?: number
  category?: 'alarm' | 'beams'
  maxLevel?: number
}

export interface RingDeviceHistoryEvent {
  msg: 'DataUpdate'
  datatype: MessageDataType
  body: any // Skipping for now
}
// ...
export interface AccountMonitoringStatus {
  accountUuid: string
  externalServiceConfigType: 'rrms' | string
  accountState: 'PROFESSIONAL' | string
  eligibleForDispatch: boolean
  addressComplete: boolean
  contactsComplete: boolean
  codewordComplete: boolean
  alarmSignalSent: boolean
  professionallyMonitored: boolean
  userAcceptDispatch: boolean
  installationDate: number
  externalId: string
  vrRequired: false
  vrUserOptIn: false
  cmsMonitoringType: 'full' | string
  dispatchSetupComplete: boolean
}

// eslint-disable-next-line no-shadow
export enum DispatchSignalType {
  Burglar = 'user-verified-burglar-xa',
  Fire = 'user-verified-fire-xa',
}

export type LocationModeInput = 'home' | 'away' | 'disarmed'
export type LocationMode = LocationModeInput | 'disabled' | 'unset'
export const disabledLocationModes: LocationMode[] = ['disabled', 'unset']

export interface LocationModeResponse {
  mode: LocationMode
  lastUpdateTimeMS: number
  securityStatus: {
    lu?: number
    md: 'none' | string
    returnTopic: string
  }
  readOnly: boolean
  notYetParticipatingInMode?: {
    deviceId: number
    deviceIdType: 'doorbot' | string
  }[]
}

export type LocationModeAction =
  | 'disableMotionDetection'
  | 'enableMotionDetection'
  | 'blockLiveViewLocally'
  | 'allowLiveViewLocally'

export interface LocationModeSetting {
  deviceId: string
  deviceIdType: 'doorbot' | string
  actions: LocationModeAction[]
}

export interface LocationModeSettings {
  disarmed: LocationModeSetting[]
  home: LocationModeSetting[]
  away: LocationModeSetting[]
}

export interface LocationModeSettingsResponse extends LocationModeSettings {
  lastUpdateTimeMS: number
}

export interface LocationModeSharing {
  sharedUsersEnabled: boolean
  lastUpdateTimeMS: number
}